feat: first example; wip: clean inl/impl files
