+++ .github/workflows/ProjectCore.yml
@@ -57,7 +57,7 @@ jobs:
         chmod +x ./llvm.sh
         sudo ./llvm.sh 17
     - name: Version
-      run: clang --version
+      run: clang++-17 --version
     - name: Building...
       run: CC=clang++-17 bazelisk build --config=default_compiler_opts //...
     - name: Testing...
