+++ MODULE.bazel
@@ -7,4 +7,4 @@ module(
     repo_name = "com_sacha_projectcore",
 )
 
-bazel_dep(name = "rules_cc", version = "0.0.4")
+bazel_dep(name = "rules_cc", version = "0.0.9")
