+++ src/ProjectCore/FMT/Context/BasicContext/BasicContext-impl.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "BasicContext.h"
+#include "BasicContextCore-impl.h"
+#include "BasicContextParse-impl.h"
+
+#include "TextPropertiesManager-impl.h"
