--- src/ProjectCore/FMT/Context/BasicContext/BasicContext-inl.h
@@ -1,14 +1,0 @@
-#pragma once
-
-#include "BasicContext.h"
-#include "BasicContextCore-inl.h"
-#include "BasicContextParse-inl.h"
-
-#include "TextPropertiesManager-inl.h"
-
-#include "ProjectCore/FMT/Buffer/BasicBufferIn/Integer-inl.h"
-#include "ProjectCore/FMT/Buffer/BasicBufferIn/String-inl.h"
-
-#include "ProjectCore/FMT/Buffer/FMTBufferIn/Integer-inl.h"
-#include "ProjectCore/FMT/Buffer/FMTBufferIn/String-inl.h"
-#include "ProjectCore/FMT/Buffer/FMTBufferIn/FromFormatData-inl.h"
