+++ src/ProjectCore/FMT/Context/BasicContext/TextPropertiesManager-impl.h
@@ -79,30 +79,6 @@ namespace ProjectCore::FMT::Detail
     }
 
     template <typename BasicContext>
-    template <typename T>
-    bool TextPropertiesManager<BasicContext>::GetColorCode(T& t)
-    {
-        static constexpr std::string_view colorCode[] = {
-            "black",
-            "red",
-            "green",
-            "yellow",
-            "blue",
-            "magenta",
-            "cyan",
-            "white",
-            "      ",
-            "default"
-        };
-
-        std::uint8_t step = static_cast<std::uint8_t>(m_Context.Format().IsEqualToForward('+') ? T::BaseBStep : T::BaseStep);
-        std::uint8_t code = static_cast<std::uint8_t>(m_Context.Format().GetWordFromList(colorCode));
-        if (code == m_Context.Format().GET_WORD_FROM_LIST_NOT_FOUND)     return false;
-        t = static_cast<T>(code + step);
-        return true;
-    }
-
-    template <typename BasicContext>
     void TextPropertiesManager<BasicContext>::ParseStyle()
     {
         if (m_Context.Format().IsEqualToForward(':')) {
