+++ src/ProjectCore/FMT/Context/FormatterContext/FormatTextProperties-inl.h
