--- src/ProjectCore/FMT/ExperimentalParser.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "Context/ParserContext/BasicParserContext.h"
-#include "Context/ParserContext/BasicParserContext-inl.h"
-
-#include "Context/ParserContext/UtilityFunctions.h"
